requirements.txt
streamlit
opencv-python
easyocr
numpy
Pillow
